fixed correct sensoor and deleted wrong root files
