+++ custom_components/ha_unifi_speedtest/sensor.py
@@ -3,6 +3,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity, DataUpdateCoordinator
 from homeassistant.helpers.typing import StateType
+from homeassistant.helpers.event import async_track_time_interval
 from datetime import timedelta
 import logging
 
@@ -32,8 +33,38 @@ async def async_setup_entry(
         _LOGGER,
         name=DOMAIN,
         update_method=async_update_data,
-        update_interval=timedelta(minutes=5)  # Adjust as needed
+        update_interval=timedelta(minutes=5)  # Check for updates every 5 minutes
     )
+
+    # Only set up automatic speed tests for traditional UniFi Controller (not UDM)
+    if api.controller_type != 'udm':
+        async def run_hourly_speedtest(now):
+            """Run speed test every hour"""
+            _LOGGER.info("Running automatic hourly speed test")
+            try:
+                await hass.async_add_executor_job(api.start_speed_test)
+                _LOGGER.info("Hourly speed test started successfully")
+            except Exception as e:
+                _LOGGER.error(f"Failed to start hourly speed test: {e}")
+
+        # Schedule the hourly speed test (runs every 1 hour)
+        async_track_time_interval(hass, run_hourly_speedtest, timedelta(hours=1))
+        _LOGGER.info("Hourly speed test scheduled for traditional UniFi Controller")
+
+        # Run an initial speed test on startup (after 1 minute delay to allow system to settle)
+        async def run_initial_speedtest(now):
+            """Run initial speed test on startup"""
+            _LOGGER.info("Running initial speed test on startup")
+            try:
+                await hass.async_add_executor_job(api.start_speed_test)
+                _LOGGER.info("Initial speed test started successfully")
+            except Exception as e:
+                _LOGGER.error(f"Failed to start initial speed test: {e}")
+
+        # Schedule initial speed test 1 minute after startup
+        async_track_time_interval(hass, run_initial_speedtest, timedelta(minutes=1))
+    else:
+        _LOGGER.info("UDM Pro detected - automatic speed tests disabled. Please set up scheduled speed tests in the UniFi Network web interface.")
 
     _LOGGER.info("Coordinator created, refreshing config entry.")
     await coordinator.async_config_entry_first_refresh()
@@ -72,8 +103,9 @@ class UniFiSpeedTestSensor(CoordinatorEntity):
 
     @property
     def unit_of_measurement(self) -> str:
+        """Return the unit of measurement following Home Assistant standards."""
         if self._data_key in ["download", "upload"]:
-            return "Mbps"
+            return "Mbit/s"  # Changed from "Mbps" to "Mbit/s" to match HA standards
         elif self._data_key == "ping":
             return "ms"
         return None
